feat: user info admin components more split and comments
